+++ examples/Smooth.hs
@@ -15,7 +15,7 @@ signal ts = map (first (utcToLocalTime utc)) (TS.toList ts)
 
 main = do
     ts <- TS.loadCSV TS.HasHeader parseISODateTime "testdata/co2.csv"
-    let y3 = 3 * 365 * 24 * 3600
+    let y3 = TS.years 3
     let xs = TS.rolling y3 (S.mean . V.fromList) ts
     toFile def "temp/co2_smooth.svg" $ do
         layout_title .= "CO2 level"